requirements.txt
hdbscan
librosa
matplotlib
numpy
sklearn
spectralcluster
tensorflow
umap-learn
pyannote.metrics
pyannote.audio